+++ JSM.Insight/Public/Set-ObjectPosition.ps1
@@ -1,5 +1,5 @@
 
-function New-Object {
+function Set-ObjectPosition {
     [CmdletBinding()]
     param (
         [string]$ID,